# Fixes #30053 - Audit page render slow when organization is set