The audit page taking long time to load when there are many audit logs and organization is set. This page fixes the issue