@hao-yu 2020-07-22
@kgaikwad: Thanks for the test. Just curious about the little performance improvement. Can you tell me what was your test result before and after the patch?

Just test in a reproducer with 135K audit entries:

Before:
[I|app|d93f0477] Started GET "/audits?page=1&per_page=20&search="
[I|app|d93f0477]   Parameters: {"page"=>"1", "per_page"=>"20", "search"=>""}
[I|app|d93f0477] Completed 200 OK in 77780ms (Views: 28.1ms | ActiveRecord: 76828.5ms)

After:
[I|app|cc1e6f17] Started GET "/audits?page=1&per_page=20&search="
[I|app|cc1e6f17]   Parameters: {"page"=>"1", "per_page"=>"20", "search"=>""}
[I|app|cc1e6f17] Completed 200 OK in 3601ms (Views: 15.6ms | ActiveRecord: 2839.3ms)