@kgaikwad 2020-07-22
> @kgaikwad: Thanks for the test. Just curious about the little performance improvement. Can you tell me what was your test result before and after the patch?

@hao-yu, thank you for sharing results. 
1.  On develop branch (With just 214 records)
- WEB
[I|app|6c3ea52b] Started GET "/audits?page=1&search=" for 127.0.0.1 at 2020-07-22 11:27:51 +0000
[I|app|0c78ebb3] Started GET "/notification_recipients" for 127.0.0.1 at 2020-07-22 11:27:51 +0000
[I|app|6c3ea52b] Processing by AuditsController#index as JSON
[I|app|6c3ea52b]   Parameters: {"page"=>"1", "search"=>""}
................
[D|sql|6c3ea52b]   ↳ app/controllers/audits_controller.rb:10:in `index'
[I|app|6c3ea52b] Completed 200 OK in 1636ms (Views: 16.8ms | ActiveRecord: 200.1ms | Allocations: 617139)

- API
[I|app|038c4f14] Started GET "/api/audits" for 127.0.0.1 at 2020-07-22 11:07:14 +0000
[I|app|038c4f14] Processing by Api::V2::AuditsController#index as JSON
[I|app|038c4f14]   Parameters: {"apiv"=>"v2"}
............
[I|app|038c4f14] Completed 200 OK in 878ms (Views: 827.9ms | ActiveRecord: 27.1ms | Allocations: 261240)

2. With pull-request changes (With just 214 records)

- WEB
[I|app|86cf2281] Processing by AuditsController#index as JSON
...........
[I|app|86cf2281] Completed 200 OK in 1690ms (Views: 30.8ms | ActiveRecord: 201.4ms | Allocations: 583445)

- API
[I|app|6d3ff47b] Processing by Api::V2::AuditsController#index as JSON
[I|app|6d3ff47b]   Parameters: {"apiv"=>"v2"}
...........
[I|app|6d3ff47b] Completed 200 OK in 759ms (Views: 516.1ms | ActiveRecord: 84.0ms | Allocations: 401884)


I also noticed that new count query takes more time than previous. 
1. With develop branch, example count query -

[D|sql|6c3ea52b]    (1.4ms)  SELECT COUNT(*) FROM "audits" WHERE ((("audits"."auditable_type" IN ($1, $2, $3, $4, $5, $6) OR "audits"."auditable_type" IN ($7, $8, $9, $10, $11, $12, $13, $14, $15, $16) AND (audits.id IN (184,116,87,71,68,51,80,70,52,190,162,84,170,176,169,92,237,69,180,97,238,236,108,156,59,197,65,98,173,149,311,200,189,239,235,161,117,88,82,188,119,240,153,147,202,196,79,210,187,85,72,95,81,244,160,232,77,302,172,198,212,242,204,249,165,151,91,209,74,179,225,138,181,245,205,96,83,67,201,63,90,105,216,174,213,107,233,86,168,163,223,93,89,241,221,167,14,231,66,158,155,118,195,207,199,75,194,243,99,152,206,157,53,230,301,183,136,150,139,193,12,164,248,137,208,78,159,94,218,186,154,234,211,177,76,106,185,64,178,148,203,58))) OR 1=0 AND "audits"."id" IN (SELECT "taxable_taxonomies"."taxable_id" FROM "taxable_taxonomies" WHERE "taxable_taxonomies"."taxonomy_id" = $17 AND "taxable_taxonomies"."taxable_type" = $18)) OR "audits"."auditable_type" = $19 AND "audits"."id" IN (SELECT "taxable_taxonomies"."taxable_id" FROM "taxable_taxonomies" WHERE "taxable_taxonomies"."taxonomy_id" = $20 AND "taxable_taxonomies"."taxable_type" = $21))  [["auditable_type", "Image"], ["auditable_type", "Operatingsystem"], ["auditable_type", "ComputeAttribute"], ["auditable_type", "Setting"], ["auditable_type", "KeyPair"], ["auditable_type", "Nic::Managed"], ["auditable_type", "Host::Base"], ["auditable_type", "ProvisioningTemplate"], ["auditable_type", "SmartProxy"], ["auditable_type", "Hostgroup"], ["auditable_type", "Filter"], ["auditable_type", "ComputeResource"], ["auditable_type", "HttpProxy"], ["auditable_type", "User"], ["auditable_type", "Parameter"], ["auditable_type", "Domain"], ["taxonomy_id", 1], ["taxable_type", "Audited::Audit"], ["auditable_type", "Organization"], ["taxonomy_id", 2], ["taxable_type", "Audited::Audit"]]

2. with pull-request change:

[D|sql|86cf2281]    (1.5ms)  SELECT COUNT(DISTINCT "audits"."id") FROM "audits" LEFT JOIN taxable_taxonomies AS tto ON audits.id = tto.taxable_id AND tto.taxable_type = 'Audited::Audit' AND tto.taxonomy_id IN (1) LEFT JOIN taxable_taxonomies AS ttl ON audits.id = ttl.taxable_id AND ttl.taxable_type = 'Audited::Audit' AND ttl.taxonomy_id IN (2) WHERE ((((("tto"."taxonomy_id" IN (1) AND "ttl"."taxonomy_id" IN (2) AND "audits"."auditable_type" IN ('Host::Base', 'ProvisioningTemplate', 'SmartProxy', 'Hostgroup', 'Filter', 'ComputeResource', 'HttpProxy', 'User', 'Parameter', 'Domain')) OR "audits"."auditable_type" IN ('Image', 'Operatingsystem', 'ComputeAttribute', 'Setting', 'KeyPair', 'Nic::Managed')) OR (1=0 AND "tto"."taxonomy_id" IN (1))) OR ("audits"."auditable_type" IN ('Organization') AND "ttl"."taxonomy_id" IN (2))))

That's why, I requested @ezr-ondrej to test and confirm.