@hao-yu 2020-07-22
@kgaikwad @ezr-ondrej Yeah. The new query is slightly slower than the old query when the records count is small (a few hundreds to a few thousands) but it is very insignificant, like < 50ms. so I think it is definitely worth.

Test for 2.4k audit entries

Before:
[I|app|50d9321b] Started GET "/audits?page=1&per_page=20&search="
[I|app|50d9321b]   Parameters: {"page"=>"1", "per_page"=>"20", "search"=>""}
[I|app|50d9321b] Completed 200 OK in 783ms (Views: 47.0ms | ActiveRecord: 92.6ms)

After:
[I|app|a399934e] Started GET "/audits?page=1&per_page=20&search="
[I|app|a399934e]   Parameters: {"page"=>"1", "per_page"=>"20", "search"=>""}
 [I|app|a399934e] Completed 200 OK in 796ms (Views: 44.6ms | ActiveRecord: 121.7ms)