@kgaikwad 2020-07-22
> @kgaikwad @ezr-ondrej Yeah. The new query is slightly slower than the old query when the records count is small (a few hundreds to a few thousands) but it is very insignificant, like < 50ms. so I think it is definitely worth.

@hao-yu - yes, definitely it is! :+1: 
I apologize for mentioning `little` based on less records.  
Let's wait for inputs from @ezr-ondrej otherwise good to go!