@hao-yu 2020-08-06
@ezr-ondrej Thanks for the test script. I am not exactly sure why the patched version would perform so badly in the test but there is an issue to the script. It is so memory hunger. I am unable to create large records for testing. It will eat all my 8GB + 1.5G swap by just creating 5 - 6K of records.

So I wrote a new rake task to run the benchmark test outside of the test framework.
**Note** I have factorybot issue when running the rake so i just comment the following out to stop the issue:
test/factories/host_related.rb
    #trait :with_build do
    #  build { true }
    #end

https://gist.github.com/hao-yu/68a4852e5541b6fb3f8d6ee5837e312b

The tests indicate that the patched version performs faster as the record count grows, especially in pagination.  See the results in:

https://pastebin.com/tNRgRJRp